requirements.txt
requests
PyYAML
arxiv
ddgs
colorlog
gradio
pandas
